arm7: merge SND_lfo into SND_exChannel
