+++ arm7/arm7.lsf
@@ -45,7 +45,6 @@ Autoload WRAM
 	Object SND_main.o
 	Object SND_capture.o
 	Object SND_channel.o
-	Object SND_lfo.o
 	Object SND_exChannel.o
 	Object SND_seq.o
 	Object SND_bank.o
