+++ arm7/lib/include/SND_lfo.h
@@ -5,9 +5,4 @@
 
 #include "nitro/SND_exChannel_shared.h"
 
-void SND_InitLfoParam(struct SNDLfoParam *lfoParam);
-void SND_StartLfo(struct SNDLfo *lfo);
-void SND_UpdateLfo(struct SNDLfo *lfo);
-int SND_GetLfoValue(struct SNDLfo *lfo);
-
 #endif //GUARD_SND_LFO_H
